[Phase 2 #104] Unify jeod_math::JeodQuat with jeod_quantities::JeodQuat
Replace the local `struct JeodQuat { data: [f64; 4] }` in
`jeod_math::quaternion` with a re-export of the canonical
`jeod_quantities::JeodQuat` (= `Quat<ScalarFirst, LeftTransform>`). The
workspace now has exactly one quaternion type; `jeod_math::JeodQuat` is a
naming alias kept for call-site ergonomics.

Changes:

- Inherent impls for `identity`, `new`, `scalar`, `vector`, `norm_sq`,
  `normalize` (JEOD fast-Padé with canonical-hemisphere flip),
  `conjugate`, `multiply`, `from_glam`/`to_glam`,
  `left_quat_from_eigen_rotation`, `left_quat_from_transformation`,
  `left_quat_to_transformation`, and `left_quat_transform` now live in
  `jeod_quantities::quat` on `Quat<ScalarFirst, LeftTransform>` (the
  orphan rule requires the defining crate for inherent impls).
- `NormalizedQuat<ScalarFirst, LeftTransform>` gains witness-gated
  `left_quat_to_transformation` and `left_quat_transform` methods that
  share a private `left_quat_to_transformation_impl` helper with the raw
  `JeodQuat` versions, so new callers can migrate to the witness type.
- `NORM_LIMIT` stays in `jeod_math::quaternion` per the phase plan; the
  constant is duplicated (private) in `jeod_quantities::quat` because
  `jeod_math` cannot be reached from below. A Tier 1 test pins the two
  copies to the same JEOD value so drift is caught immediately.
- `ScalarFirst`, `ScalarLast`, `LeftTransform`, `RightTransform` now
  derive `PartialEq`, `Eq`, `Hash` so downstream structs holding a
  `JeodQuat` can still `#[derive(PartialEq)]` unchanged.
- `jeod_math::lib.rs` now re-exports `jeod_quantities::prelude::*` plus
  the explicit quaternion-convention tag set, matching the Phase 2 plan.
- One internal struct-literal use (`JeodQuat { data }` in
  `euler_angles.rs`) is rewritten to `JeodQuat::from_array(data)` since
  the unified type's phantom fields are private.

Existing call sites (`JeodQuat::new`, `.data`, `.scalar()`, `.vector()`,
`.multiply()`, `.normalize()`, `.left_quat_to_transformation()`, etc.)
continue to work unchanged — this commit intentionally preserves every
public method name and signature so no downstream crate needs editing.

Verification:
- fmt, clippy (workspace + tests) clean.
- `cargo nextest run -p jeod_math`: 74/74 passing.
- `cargo nextest run --workspace -E 'not test(tier3_)'`: 628/628 passing.
- `cargo nextest run -p jeod_runner -E 'test(tier3_simulation_run2)'`:
  3/3 passing (tier 3 canary unchanged numerically).
