+++ crates/jeod_math/src/lib.rs
@@ -1,3 +1,13 @@
+//! `jeod_math` — JEOD-faithful math kernels.
+//!
+//! Phase 2 of the type-system refactor (#104) unifies the quaternion type
+//! with [`jeod_quantities`]. `JeodQuat` is now a re-export of
+//! `jeod_quantities::JeodQuat` (the canonical `Quat<ScalarFirst,
+//! LeftTransform>` type alias), and all algebraic/conversion methods live
+//! on that unified type so there is only one quaternion in the workspace.
+
+pub use jeod_quantities::prelude::*;
+
 pub mod error;
 pub mod euler_angles;
 pub mod geodetic;
@@ -15,8 +25,11 @@ pub use euler_angles::{
     compute_quaternion_from_euler_angles, EulerSequence, ALL_SEQUENCES,
 };
 pub use geodetic::{cartesian_to_geodetic, geodetic_to_cartesian, GeodeticState};
+pub use jeod_quantities::{
+    JeodQuat, Layout, LeftTransform, NormalizedQuat, Quat, RightTransform, ScalarFirst, ScalarLast,
+    Transform,
+};
 pub use lvlh::{compute_lvlh_frame, LvlhFrame};
 pub use orbital_elements::OrbitalElements;
-pub use quaternion::JeodQuat;
 pub use solar_beta::solar_beta_angle;
 pub use types::*;
